Rename auth/ to identity/
