+++ src/App.tsx
@@ -11,8 +11,8 @@ import CliInstructions from "./components/transferData/CliInstructions";
 import TemplatesPage from "./components/templates/TemplatesPage";
 import PrivacyAndSecurityPage from "./components/privacyAndSecurity/PrivacyAndSecurityPage";
 import UserAccountPage from "./components/userAccount/UserAccountPage";
-import Register from "./components/register/Register";
-import Unactivated from "./components/register/Unactivated";
+import Register from "./auth/Register";
+import Unactivated from "./auth/Unactivated";
 import history from "./auth/History";
 import AssayInstructions from "./components/transferData/AssayInstructions";
 import AuthProvider from "./auth/AuthProvider";
